requirements.txt
# Existing
kafka-python==2.0.2
<<<<<<< HEAD
newsapi==1.4
yfinance==0.2.40
pandas==2.1.0
numpy==1.25.0
=======
newsapi-python
yfinance==0.2.32
pandas==2.0.3
numpy==1.24.3
>>>>>>> 1d33bbb6974900a380768e2397c8086468d74ab2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
ta==0.11.0
scipy==1.11.4

# NEW for Week 3: ML Libraries
xgboost==2.0.0
scikit-learn==1.3.2
lightgbm==4.1.0
optuna==3.14.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
